Revert "(#7) Harden parsers against brittle table assumptions"
This reverts commit 341e86e1023f85abe060cb8d4b11e0903f193c5c.
